+++ headers-server.js
@@ -16,22 +16,6 @@ function ipChain(headers, connection) {
 }
 
 /*
- * The client will request this "script", and send a unique token with it,
- * which we later use to re-associate the headers from this request with
- * the user's livedata session (since XHR requests only send a subset of
- * all the regular headers).
- */
-WebApp.connectHandlers.use('/headersHelper.js', function(req, res, next) {
-  var token = req.query.token;
-
-  req.headers['x-ip-chain'] = ipChain(req.headers, req.connection).join(',');
-  headers.list[token] = req.headers;
-
-  res.writeHead(200, { 'Content-type': 'application/javascript' });
-  res.end("Package.headers.headers.store(" + JSON.stringify(req.headers) + ");", 'utf8');
-});
-
-/*
  * After user has requested the headers (which were stored in headers.list
  * at the same time with the client's token, the below is called, which we
  * use to re-associate with the user's livedata session (see above)
@@ -42,7 +26,12 @@ Meteor.methods({
       var data = this.connection || this._sessionData;
       data.headers = headers.list[token];
       headerDep(data).changed();
-      delete headers.list[token];
+
+      // Don't do this until Meteor resumes sessions.  Consider
+      // longer cleanup time, and keeping last reassocation time.
+      // Or on disconnect, put back in the list with disconnect
+      // time and keep that for cleanup_time (can do in 0.7+).
+      // delete headers.list[token];
     }
   }
 });
@@ -64,14 +53,24 @@ function headerDep(obj) {
     obj.headerDep = new Deps.Dependency();
   return obj.headerDep;
 }
+ 
+/*
+ * Provide helpful hints for incorrect usage
+ */
+function checkSelf(self, funcName) {
+  if (!self || (!self.connection && !self._session && !self._sessionData))
+    throw new Error('Call headers.' + funcName + '(this) only from within a '
+    	+ 'method or publish function.  With callbacks / anonymous '
+    	+ 'functions, use: var self=this; and call headers.'+funcName+'(self);');
+}
 
 /*
  * Usage in a Meteor method/publish: headers.get(this, 'host')
  */
 headers.get = function(self, key) {
-  if (!self)
-    throw new Error('Must be called like this on the server: headers.get(this)');
+  checkSelf(self, 'get');
   var sessionData = self.connection || (self._session ? self._session.sessionData : self._sessionData);
+
   headerDep(sessionData).depend();
   if (!(sessionData && sessionData.headers))
     return key ? undefined : {};
@@ -79,16 +78,14 @@ headers.get = function(self, key) {
 }
 
 headers.ready = function(self) {
-  if (!self)
-    throw new Error('Must be called like this on the server: headers.get(this)');
+  checkSelf(self, 'ready');
   var sessionData = self.connection || (self._session ? self._session.sessionData : self._sessionData);
   headerDep(sessionData).depend();
   return Object.keys(sessionData.headers).length > 0;
 }
 
 headers.getClientIP = function(self, proxyCount) {
-  if (!self)
-    throw new Error('Must be called like this on the server: headers.getClientIP(this)');
+  checkSelf(self, 'getClientIP');
   var chain = this.get(self, 'x-ip-chain').split(',');
   if (typeof(proxyCount) == 'undefined')
     proxyCount = this.proxyCount;
@@ -96,15 +93,16 @@ headers.getClientIP = function(self, proxyCount) {
 }
 
 /*
- * Get the IP for the livedata connection used by a Method (see README.md)
+ * Retrieve header(s) for the current method socket (see README.md)
  */
-headers.methodClientIP = function(self, proxyCount) {
+headers.methodGet = function(self, header) {
   var session, headers, chain;
- 
+  checkSelf(self, 'methodGet');
+
   if (self.connection) {
     // Meteor 0.6.7+
     session = Meteor.server.sessions[self.connection.id];
-  } else {
+  } else if (self._session || self._sessionData) {
     // convoluted way to find our session in Meteor < 0.6.7
     var sessionData = self._session ? self._session.sessionData : self._sessionData;
     var token, session;
@@ -116,9 +114,103 @@ headers.methodClientIP = function(self, proxyCount) {
   }
 
   headers = session.socket.headers;
-  chain = ipChain(headers, session.socket);
+  if (!headers['x-ip-chain'])
+	  headers['x-ip-chain'] = ipChain(headers, session.socket);
+
+  return header ? headers[header] : headers;
+}
+
+/*
+ * Get the IP for the livedata connection used by a Method (see README.md)
+ */
+headers.methodClientIP = function(self, proxyCount) {
+  checkSelf(self, 'methodClientIP');
+  var chain = this.methodGet(self, 'x-ip-chain');
   if (typeof(proxyCount) == 'undefined')
     proxyCount = this.proxyCount;
   return chain[chain.length - proxyCount - 1];
 }
 
+/*
+ * All this code is below adapted from Fast Render by Arunoda Susiripala
+ * https://github.com/arunoda/meteor-fast-render/blob/master/lib/server/inject.js
+ *
+ */
+
+//When a HTTP Request comes, we need to figure out is it a proper request
+//then get some query data
+//then hijack html return by meteor
+//code below, does that in abstract way
+
+var http = Npm.require('http');
+
+var originalWrite = http.OutgoingMessage.prototype.write;
+http.OutgoingMessage.prototype.write = function(chunk, encoding) {
+  //prevent hijacking other http requests
+  if(this.mhData && !this.injected && 
+    encoding === undefined && /<!DOCTYPE html>/.test(chunk)) {
+
+    //if cors headers included if may cause some security holes. see more: 
+    //so we simply turn off fast-render if we detect an cors header
+    //read more: http://goo.gl/eGwb4e
+    if(this._headers['access-control-allow-origin']) {
+      var wanrMessage = 
+        'warn: fast-render turned off due to CORS headers. read more: http://goo.gl/eGwb4e';
+      console.warn(wanrMessage);
+      originalWrite.call(this, chunk, encoding);
+      return;
+    }
+
+    // Indentation same as __meteor_runtime_config__
+    var injectHtml = '<script type="text/javascript">__headers__ = \n'
+    	+ JSON.stringify(this.mhData) + '</script>\n';
+
+	chunk = chunk.replace('<head>', '<head>\n' + injectHtml + '\n');
+    this.injected = true;
+  }
+
+  originalWrite.call(this, chunk, encoding);
+};
+
+//meteor algorithm to check if this is a meteor serving http request or not
+//add routepolicy package to the fast-render
+function appUrl(url) {
+  if (url === '/favicon.ico' || url === '/robots.txt')
+    return false;
+
+  // NOTE: app.manifest is not a web standard like favicon.ico and
+  // robots.txt. It is a file name we have chosen to use for HTML5
+  // appcache URLs. It is included here to prevent using an appcache
+  // then removing it from poisoning an app permanently. Eventually,
+  // once we have server side routing, this won't be needed as
+  // unknown URLs with return a 404 automatically.
+  if (url === '/app.manifest')
+    return false;
+
+  // Avoid serving app HTML for declared routes such as /sockjs/.
+  if (typeof(RoutePolicy) != 'undefined' && RoutePolicy.classify(url))
+    return false;
+
+  // we currently return app HTML on all URLs by default
+  return true;
+};
+
+// Check page and add mhData if relevant
+WebApp.connectHandlers.use(Npm.require('connect').cookieParser());
+WebApp.connectHandlers.use(function(req, res, next) {
+  if(appUrl(req.url)) {
+  	// create a unique token to store headers for this request (see README)
+  	var token = new Date().getTime() + Math.random();
+  	headers.list[token] = req.headers;
+
+  	// for the getClientIP helper
+    req.headers['x-ip-chain'] = ipChain(req.headers, req.connection).join(',');
+
+    // data to be injected into initial page HEAD (see http.outgoing... above)
+    res.mhData = { headers: req.headers, token: token };
+
+	next();
+  } else {
+    next();
+  }
+});